Cross-build jawn for Scala.js
