+++ modules/jawn/js/src/main/scala/io/circe/jawn/JawnParserPlatform.scala
@@ -1,0 +1,3 @@
+package io.circe.jawn
+
+private[jawn] trait JawnParserPlatform
